# Optimize DataPageCompress: