- Pre-allocate more buffers
- Reuse one buffer