@alex 2020-05-10
Benchmark from #257:

```
name         old time/op    new time/op    delta
WriteCSV-44    10.2ms ± 7%     9.6ms ± 3%   -6.27%  (p=0.000 n=19+20)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    10.7MB ± 0%     9.1MB ± 0%  -14.70%  (p=0.000 n=20+19)

name         old allocs/op  new allocs/op  delta
WriteCSV-44      103k ± 0%      103k ± 0%   -0.37%  (p=0.000 n=20+17)
```